+++ api/core/v2/check.go
@@ -49,6 +49,17 @@ const (
 	// RegistrationCheckName is the name of the check that is created when an
 	// entity sends a keepalive and the entity does not yet exist in the store.
 	RegistrationCheckName = "registration"
+
+	// MemoryScheduler indicates that a check is scheduled in-memory.
+	MemoryScheduler = "memory"
+
+	// EtcdScheduler indicates that a check is scheduled with etcd leases and
+	// watchers.
+	EtcdScheduler = "etcd"
+
+	// PostgresScheduler indicates that a check is scheduled with postgresql,
+	// using transactions and asynchronous notification (NOTIFY).
+	PostgresScheduler = "postgres"
 )
 
 // OutputMetricFormats represents all the accepted output_metric_format's a check can have
@@ -113,6 +124,7 @@ func NewCheck(c *CheckConfig) *Check {
 		EnvVars:              c.EnvVars,
 		DiscardOutput:        c.DiscardOutput,
 		MaxOutputSize:        c.MaxOutputSize,
+		Scheduler:            c.Scheduler,
 	}
 	if check.Labels == nil {
 		check.Labels = make(map[string]string)